chore: move function and interface
